+++ src/main/java/com/example/concerts/ConcertApiApplication.java
@@ -1,4 +1,4 @@
-package com.example.concertapi;
+package com.example.concerts;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
