+++ src/main/java/com/example/concerts/concert/ConcertController.java
@@ -1,4 +1,4 @@
-package com.example.concertapi.concert;
+package com.example.concerts.concert;
 
 import org.springframework.web.bind.annotation.*;
