+++ src/main/java/com/example/concerts/concert/ConcertDTO.java
@@ -1,4 +1,4 @@
-package com.example.concertapi.concert;
+package com.example.concerts.concert;
 
 import java.time.LocalDateTime;
