+++ src/main/java/com/example/concerts/concert/ConcertRepository.java
@@ -1,4 +1,4 @@
-package com.example.concertapi.concert;
+package com.example.concerts.concert;
 
 import org.springframework.data.jpa.repository.JpaRepository;
